refactor(daemon): delete the unused EventLogger module (#1165 finding 7)
`EventLogger::new` and `log_event` had zero production callers -- every
construction was in the module's own tests. `daemon.log` was therefore an
observability surface that never received a line, which is worse than a missing
one: it reads as covered.

Delete `daemon/event_log.rs` (582 lines: EventLogger, CompileEvent,
CompileOutcome, LogWriter, their rotation/GC machinery and ~11 tests) and drop
the `pub use event_log::EventLogger` re-export.

Its only live parts were two helpers used by the compile journal --
`open_append` (adds FILE_SHARE_DELETE on Windows) and `format_timestamp` with
its `civil_from_days` -- so they move to `compile_journal/log_io.rs`, next to
their sole consumer, with the timestamp test that covered them.

Stale references updated rather than left dangling: the `daemon.log.*` sweep in
`cli/runtime.rs` now says those archives are legacy and nothing writes them; the
two "same pattern as EventLogger" module docs and the matching line in
DESIGN_DECISIONS.md no longer point at a type that does not exist.

Note this removes a `pub` item from a published crate
(`zccache_daemon_core::daemon::EventLogger`). It is dead code with no in-tree
callers, and the issue recommends deletion, but the semver visibility is worth
a maintainer's eye.
